refactor : 의존성을 고려해 패키지 위치 변경
