+++ src/main/java/kr/co/conceptbe/idea/application/IdeaService.java
@@ -8,9 +8,8 @@ import java.util.stream.Collectors;
 import kr.co.conceptbe.auth.presentation.dto.AuthCredentials;
 import kr.co.conceptbe.bookmark.Bookmark;
 import kr.co.conceptbe.branch.domain.persistense.BranchRepository;
-import kr.co.conceptbe.idea.presentation.dto.response.FindIdeaWriteResponse;
+import kr.co.conceptbe.idea.application.response.FindIdeaWriteResponse;
 import kr.co.conceptbe.member.exception.UnAuthorizedMemberException;
-import kr.co.conceptbe.member.exception.NotFoundMemberException;
 import kr.co.conceptbe.purpose.domain.persistence.PurposeRepository;
 import kr.co.conceptbe.region.domain.presentation.RegionRepository;
 import kr.co.conceptbe.teamrecruitment.domain.persistence.TeamRecruitmentCategoryRepository;
@@ -22,9 +21,9 @@ import kr.co.conceptbe.idea.domain.IdeaLike;
 import kr.co.conceptbe.idea.domain.persistence.IdeaLikesRepository;
 import kr.co.conceptbe.idea.domain.persistence.IdeaRepository;
 import kr.co.conceptbe.idea.dto.IdeaDetailResponse;
-import kr.co.conceptbe.idea.presentation.dto.response.BestIdeaResponse;
-import kr.co.conceptbe.idea.presentation.dto.request.IdeaRequest;
-import kr.co.conceptbe.idea.presentation.dto.response.IdeaResponse;
+import kr.co.conceptbe.idea.application.response.BestIdeaResponse;
+import kr.co.conceptbe.idea.application.request.IdeaRequest;
+import kr.co.conceptbe.idea.application.response.IdeaResponse;
 import kr.co.conceptbe.member.domain.Member;
 import kr.co.conceptbe.member.persistence.MemberRepository;
 import lombok.RequiredArgsConstructor;
