+++ src/main/java/kr/co/conceptbe/idea/presentation/IdeaController.java
@@ -7,10 +7,10 @@ import kr.co.conceptbe.common.auth.Auth;
 import kr.co.conceptbe.common.auth.OptionalAuth;
 import kr.co.conceptbe.idea.application.IdeaService;
 import kr.co.conceptbe.idea.dto.IdeaDetailResponse;
-import kr.co.conceptbe.idea.presentation.dto.request.IdeaRequest;
-import kr.co.conceptbe.idea.presentation.dto.response.BestIdeaResponse;
-import kr.co.conceptbe.idea.presentation.dto.response.FindIdeaWriteResponse;
-import kr.co.conceptbe.idea.presentation.dto.response.IdeaResponse;
+import kr.co.conceptbe.idea.application.request.IdeaRequest;
+import kr.co.conceptbe.idea.application.response.BestIdeaResponse;
+import kr.co.conceptbe.idea.application.response.FindIdeaWriteResponse;
+import kr.co.conceptbe.idea.application.response.IdeaResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
